# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Money Blog API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Category {
  id: number;
  slug: string;
  name: string;
  description: string;
  color: string;
  icon: string;
  postCount: number;
}

export type PostSummaryDifficulty =
  (typeof PostSummaryDifficulty)[keyof typeof PostSummaryDifficulty];

export const PostSummaryDifficulty = {
  beginner: "beginner",
  intermediate: "intermediate",
  advanced: "advanced",
} as const;

export type PostSummaryIncomeType =
  (typeof PostSummaryIncomeType)[keyof typeof PostSummaryIncomeType];

export const PostSummaryIncomeType = {
  passive: "passive",
  active: "active",
  hybrid: "hybrid",
} as const;

export interface Author {
  name: string;
  role: string;
  avatar: string;
}

export interface PostSummary {
  id: number;
  slug: string;
  title: string;
  excerpt: string;
  coverImage: string;
  category: Category;
  difficulty: PostSummaryDifficulty;
  incomeType: PostSummaryIncomeType;
  tags: string[];
  premium: boolean;
  readMinutes: number;
  views: number;
  likes: number;
  author: Author;
  publishedAt: string;
}

export type Post = PostSummary & {
  /** Markdown body */
  content: string;
};

export interface PostList {
  items: PostSummary[];
  total: number;
  limit: number;
  offset: number;
}

export type ResourcePricing =
  (typeof ResourcePricing)[keyof typeof ResourcePricing];

export const ResourcePricing = {
  free: "free",
  freemium: "freemium",
  paid: "paid",
} as const;

export interface Resource {
  id: number;
  title: string;
  description: string;
  url: string;
  category: string;
  pricing: ResourcePricing;
  tags: string[];
  logo: string;
}

export interface TagCount {
  tag: string;
  count: number;
}

export interface SiteStats {
  totalPosts: number;
  totalCategories: number;
  totalResources: number;
  totalReads: number;
  totalSubscribers: number;
}

export interface CountResponse {
  count: number;
}

export interface PostSummaryAi {
  tldr: string;
  keyTakeaways: string[];
}

export type ChatTurnRole = (typeof ChatTurnRole)[keyof typeof ChatTurnRole];

export const ChatTurnRole = {
  user: "user",
  assistant: "assistant",
} as const;

export interface ChatTurn {
  role: ChatTurnRole;
  content: string;
}

export interface ChatRequest {
  message: string;
  history?: ChatTurn[];
}

export interface ChatResponse {
  reply: string;
  suggestions: string[];
}

export interface NewsletterRequest {
  email: string;
}

export interface NewsletterResponse {
  ok: boolean;
  message: string;
}

export type ListPostsParams = {
  /**
   * Category slug
   */
  category?: string;
  difficulty?: ListPostsDifficulty;
  incomeType?: ListPostsIncomeType;
  tag?: string;
  search?: string;
  premium?: boolean;
  sort?: ListPostsSort;
  /**
   * @maximum 50
   */
  limit?: number;
  offset?: number;
};

export type ListPostsDifficulty =
  (typeof ListPostsDifficulty)[keyof typeof ListPostsDifficulty];

export const ListPostsDifficulty = {
  beginner: "beginner",
  intermediate: "intermediate",
  advanced: "advanced",
} as const;

export type ListPostsIncomeType =
  (typeof ListPostsIncomeType)[keyof typeof ListPostsIncomeType];

export const ListPostsIncomeType = {
  passive: "passive",
  active: "active",
  hybrid: "hybrid",
} as const;

export type ListPostsSort = (typeof ListPostsSort)[keyof typeof ListPostsSort];

export const ListPostsSort = {
  recent: "recent",
  popular: "popular",
  trending: "trending",
} as const;

export type ListTrendingPostsParams = {
  limit?: number;
};

export type ListResourcesParams = {
  category?: string;
  pricing?: ListResourcesPricing;
  search?: string;
  /**
   * @maximum 60
   */
  limit?: number;
};

export type ListResourcesPricing =
  (typeof ListResourcesPricing)[keyof typeof ListResourcesPricing];

export const ListResourcesPricing = {
  free: "free",
  freemium: "freemium",
  paid: "paid",
} as const;

export type ListPopularTagsParams = {
  limit?: number;
};