# lib/api-zod/src/generated/types/chatRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Money Blog API
 * OpenAPI spec version: 0.1.0
 */
import type { ChatTurn } from "./chatTurn";

export interface ChatRequest {
  message: string;
  history?: ChatTurn[];
}