# lib/api-zod/src/generated/types/resource.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Money Blog API
 * OpenAPI spec version: 0.1.0
 */
import type { ResourcePricing } from "./resourcePricing";

export interface Resource {
  id: number;
  title: string;
  description: string;
  url: string;
  category: string;
  pricing: ResourcePricing;
  tags: string[];
  logo: string;
}